nachok

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -2,15 +2,15 @@
 
 int main(void)
 {
-    int n = 50; // Number of Fibonacci numbers to generate
-    unsigned long long int fib1 = 1, fib2 = 2, nextFib;
+    int n = 50;
+    unsigned long fib1 = 1, fib2 = 2, nextFib;
 
-    printf("%llu, %llu", fib1, fib2);
+    printf("%lu, %lu", fib1, fib2);
 
     for (int i = 3; i <= n; i++)
     {
         nextFib = fib1 + fib2;
-        printf(", %llu", nextFib);
+        printf(", %lu", nextFib);
         fib1 = fib2;
         fib2 = nextFib;
     }
